refactor: 디렉토리 구조 수정(#4)
