+++ src/main/java/project/backend/domain/auth/repository/RefreshTokenRedisRepository.java
@@ -1,0 +1,9 @@
+package project.backend.domain.auth.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import project.backend.global.auth.token.RefreshToken;
+
+public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, Long> {
+
+  RefreshToken findByRefreshToken(String refreshToken);
+}
