+++ src/main/java/project/backend/domain/user/entity/User.java
@@ -1,0 +1,44 @@
+package project.backend.domain.user.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import project.backend.global.common.BaseEntity;
+
+@Entity
+@Getter
+@Table(name = "users")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "user_id")
+  private Long id;
+
+  @Column(name = "user_name")
+  private String name;
+
+  @Column(name = "user_email")
+  private String email;
+
+  @Builder
+  private User(String name, String email) {
+    this.name = name;
+    this.email = email;
+  }
+
+  public static User createUser(String email, String name) {
+    return User.builder()
+               .email(email)
+               .name(name)
+               .build();
+  }
+}
