+++ src/main/java/project/backend/domain/user/repositoy/UserRepository.java
@@ -1,0 +1,12 @@
+package project.backend.domain.user.repositoy;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import project.backend.domain.user.entity.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+  Optional<User> findByEmail(String email);
+}
