+++ src/main/java/project/backend/global/auth/oauth/KakaoUserDetails.java
@@ -1,0 +1,34 @@
+package project.backend.global.auth.oauth;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+@Getter
+@RequiredArgsConstructor
+public class KakaoUserDetails implements OAuth2User {
+
+  private final Long id;
+  private final String email;
+  private final List<? extends GrantedAuthority> authorities;
+  private final Map<String, Object> attributes;
+
+  @Override
+  public String getName() {
+    return email;
+  }
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    return authorities;
+  }
+
+  @Override
+  public Map<String, Object> getAttributes() {
+    return attributes;
+  }
+}
