+++ src/main/java/project/backend/global/auth/token/TokenResponse.java
@@ -1,0 +1,16 @@
+package project.backend.global.auth.token;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class TokenResponse {
+
+  private final String accessToken;
+  private final String refreshToken;
+
+  public static TokenResponse of(final String accessToken, final String refreshToken) {
+    return new TokenResponse(accessToken, refreshToken);
+  }
+}
