--- src/main/java/project/backend/user/infra/security/jwt/token/TokenResponse.java
@@ -1,16 +1,0 @@
-package project.backend.user.infra.security.jwt.token;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class TokenResponse {
-
-  private final String accessToken;
-  private final String refreshToken;
-
-  public static TokenResponse of(final String accessToken, final String refreshToken) {
-    return new TokenResponse(accessToken, refreshToken);
-  }
-}
