--- src/main/java/project/backend/user/repositoy/RefreshTokenRedisRepository.java
@@ -1,9 +1,0 @@
-package project.backend.user.repositoy;
-
-import org.springframework.data.repository.CrudRepository;
-import project.backend.user.infra.security.jwt.token.RefreshToken;
-
-public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, Long> {
-
-  RefreshToken findByRefreshToken(String refreshToken);
-}
