--- src/main/java/project/backend/user/repositoy/UserRepository.java
@@ -1,12 +1,0 @@
-package project.backend.user.repositoy;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import project.backend.user.domain.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByEmail(String email);
-}
